Pdf preview: rename tests
